+++ lib/plugin-directory-utils.rb
