+++ plugins.json
