+++ sketch-slazzer.sketchplugin/Contents/Resources/icon.png
