--- sketch-slazzer/sketch-slazzer.sketchplugin/Contents/Sketch/manifest.json
@@ -1,30 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/sketch-hq/SketchAPI/develop/docs/sketch-plugin-manifest-schema.json",
-  "icon": "icon.png",
-  "commands": [
-    {
-      "name": "Remove Background",
-      "identifier": "sketch-slazzer.slazzer_cmd-identifier",
-      "script": "__slazzer_cmd.js"
-    },
-    {
-      "name": "Set API key",
-      "identifier": "slazzer.api-key",
-      "script": "__set-api-key.js"
-    }
-  ],
-  "menu": {
-    "title": "Slazzer",
-    "items": [
-      "sketch-slazzer.slazzer_cmd-identifier",
-      "slazzer.api-key"
-    ]
-  },
-  "version": "1.0.2",
-  "description": "Slazzer sketch plugin for remove image background automatically",
-  "name": "sketch-slazzer",
-  "identifier": "sketch-slazzer",
-  "disableCocoaScriptPreprocessor": true,
-  "appcast": "https://raw.githubusercontent.com/slazzercom/slazzer-sketch/master/.appcast.xml",
-  "author": "Slazzer"
-}
